Various important changes
WARNING: this compiles and runs. However, the output miniSampleOutput.kml is NOT correct (there's an exception thrown but logged).
NOTE TO SELF: this is why Scala is not a more popular language! It won't even tell me what the three warnings are!!
NamedFunction: added ^+ method which prefixes the given argument to the existing string;
KML: 
* DANGER of infinite recursion of definition of implicit renders: def replaced with lazy val where necessary and simply val otherwise.
* use ^+ style for at least some of the renders.
Renderers: added trace messages to renderer1, etc.